object_detection_framework/requirements_virtualenv.txt
absl-py==0.8.0
alabaster==0.7.12
ascii-graph==1.5.1
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
audioread==2.1.8
Babel==2.7.0
backcall==0.1.0
beautifulsoup4==4.8.0
bleach==3.1.0
boto3==1.9.224
botocore==1.12.224
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
codecov==2.0.15
coverage==4.5.4
cryptography==2.7
cycler==0.10.0
cymem==2.0.2
Cython==0.28.4
cytoolz==0.9.0.1
DataProperty==0.43.1
decorator==4.4.0
defusedxml==0.6.0
dill==0.2.9
docutils==0.15.2
entrypoints==0.3
filelock==3.0.12
flake8==3.7.8
Flask==1.1.1
future==0.17.1
glob2==0.7
grpcio==1.23.0
h5py==2.10.0
html2text==2019.8.11
hypothesis==4.35.0
idna==2.8
imageio==2.5.0
imagesize==1.1.0
importlib-metadata==0.20
inflect==2.1.0
ipdb==0.12.2
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
json5==0.8.5
jsonschema==3.0.2
jupyter-client==5.3.1
jupyter-core==4.5.0
jupyter-tensorboard==0.1.10
jupytext==1.2.3
kiwisolver==1.1.0
libarchive-c==2.8
librosa==0.6.3
lief==0.9.0
llvmlite==0.28.0
lmdb==0.97
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mbstrdecoder==0.8.1
mccabe==0.6.1
mistune==0.8.4
mlperf-compliance==0.0.10
mock==3.0.5
more-itertools==7.2.0
msgfy==0.0.7
msgpack==0.6.1
msgpack-numpy==0.4.3.2
murmurhash==1.0.2
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.0
nltk==3.4.5
notebook==6.0.1
numba==0.43.1
numpy==1.16.4
onnx==1.5.0
opencv-python==3.4.1.15
packaging==19.1
pandas==0.24.2
pandocfilters==1.4.2
parso==0.5.1
pathvalidate==0.29.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==5.3.0
Pillow-SIMD==5.3.0.post1
pkginfo==1.5.0.1
plac==0.9.6
pluggy==0.12.0
preshed==2.0.1
progressbar==2.5
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.9.1
psutil==5.6.3
ptyprocess==0.6.0
py==1.8.0
pybind11==2.3.0
pycocotools==2.0.0
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pydot==1.4.1
pyflakes==2.1.1
Pygments==2.4.2
pynvml==8.0.4
pyOpenSSL==19.0.0
pyparsing==2.4.2
pyrsistent==0.15.4
PySocks==1.7.0
pytablewriter==0.46.1
pytest==5.1.2
pytest-cov==2.7.1
pytest-pythonpath==0.7.3
python-dateutil==2.8.0
python-nvd3==0.15.0
python-slugify==3.0.3
pytz==2019.2
PyWavelets==1.0.3
PyYAML==5.1.2
pyzmq==18.1.0
regex==2018.1.10
requests==2.22.0
resampy==0.2.2
revtok==0.0.3
ruamel.yaml==0.15.46
s3transfer==0.2.1
sacrebleu==1.2.10
sacremoses==0.0.19
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.3.1
Send2Trash==1.5.0
six==1.12.0
snowballstemmer==1.9.0
SoundFile==0.10.2
soupsieve==1.9.2
sox==1.3.7
spacy==2.0.16
Sphinx==2.2.0
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
statsmodels==0.11.1
subword-nmt==0.3.3
tabledata==0.9.1
tabulate==0.8.3
tensorboard==1.14.0
tensorboardX==2.0
terminado==0.8.2
testpath==0.4.2
text-unidecode==1.2
thinc==6.12.1
toml==0.10.0
toolz==0.10.0
torch==1.2.0
torchvision==0.4.0
tornado==6.0.3
tqdm==4.31.1
traitlets==4.3.2
typepy==0.6.0
typing==3.7.4.1
typing-extensions==3.7.4
ujson==1.35
Unidecode==1.1.1
urllib3==1.24.2
visdom==0.1.8.9
Vizer==0.1.5
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.6
wrapt==1.10.11
yacs==0.1.6
zipp==0.6.0
